% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_to_date}
\alias{char_to_date}
\title{Convert a character to a date}
\usage{
char_to_date(date, date_pattern = "\%Y-\%m-\%d")
}
\arguments{
\item{date}{A character vector of dates}

\item{date_pattern}{The pattern of the date}
}
\value{
A date vector
}
\description{
Convert a character to a date
}
\examples{
Pedixplorer:::char_to_date("2020-01-01", "\%Y-\%m-\%d")
Pedixplorer:::char_to_date("01/01/20", "\%d/\%m/\%y")
}
\keyword{internal}
