% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_color_picker.R
\name{color_picker_ui}
\alias{color_picker_ui}
\alias{color_picker_server}
\alias{color_picker_demo}
\title{Shiny modules to select colours}
\usage{
color_picker_ui(id)

color_picker_server(id, colors = NULL)

color_picker_demo()
}
\arguments{
\item{id}{A string to identify the module.}

\item{colors}{A list of variables and their default colours.}
}
\value{
A reactive list with the selected colours.
}
\description{
This function allows to select different colours for an
array of variables.
}
\examples{
if (interactive()) {
    color_picker_demo()
}
}
\keyword{internal}
