% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complete_twins}
\alias{complete_twins}
\title{Complete missing twins relationship}
\usage{
complete_twins(rel_df, multi_code = "error")
}
\arguments{
\item{rel_df}{A dataframe of relationships}

\item{multi_code}{How to handle multiple relationship codes in the same group
If "error", an error is thrown. If "warn", a warning is thrown and the
relationship code is set to twins of unknow zigosity. Default is "error".}
}
\value{
The completed dataframe of relationships
}
\description{
Given a dataframe of relationships, complete the missing twins
relationship.
}
\examples{
data(relped)
Pedixplorer:::complete_twins(relped)
}
\keyword{internal}
