% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_col_sel.R
\name{data_col_sel_demo}
\alias{data_col_sel_demo}
\title{Demo function of selecting columns module}
\usage{
data_col_sel_demo()
}
\description{
Demo function of selecting columns module
}
\examples{
data_col_sel_demo()
}
