% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_col_sel.R
\name{data_col_sel_server}
\alias{data_col_sel_server}
\title{Server function of selecting columns module}
\usage{
data_col_sel_server(id, df, cols_needed, cols_supl, title, na_omit = FALSE)
}
\arguments{
\item{id}{A string to identify the module.}

\item{df}{A reactive dataframe.}

\item{cols_needed}{A character vector of the mandatory columns.}

\item{cols_supl}{A character vector of the optional columns.}

\item{title}{A string to display in the selectInput.}

\item{na_omit}{A boolean to allow or not the selection of NA.}
}
\value{
A reactive dataframe with the selected columns renamed
to the names of cols_needed and cols_supl.
}
\description{
Server function of selecting columns module
}
\examples{
data_col_sel_demo()
}
