% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_col_sel.R
\name{data_col_sel_ui}
\alias{data_col_sel_ui}
\title{User interface of selecting columns module}
\usage{
data_col_sel_ui(id)
}
\arguments{
\item{id}{A string to identify the module.}
}
\value{
A Shiny module UI.
}
\description{
User interface of selecting columns module
}
\examples{
data_col_sel_demo()
}
