% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_import.R
\name{data_import_server}
\alias{data_import_server}
\title{Data import server}
\usage{
data_import_server(id, label = "Select data file", max_request_size = 30)
}
\arguments{
\item{id}{A string.}
}
\value{
A Shiny server.
}
\description{
R Shiny module server to import data files
}
\details{
This module allow to import multiple type of data.
The file type currently supported are csv, txt, xls, xslx, rda.
The server dynamically create a selection input if multiple
dataframe are present in the file selected.
}
\examples{
\\dontrun{
    data_import_demo()
}
}
\keyword{data}
