% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_col_sel.R
\name{distribute_by}
\alias{distribute_by}
\title{Distribute elements by group}
\usage{
distribute_by(nb_group, nb_elem, by_row = FALSE)
}
\arguments{
\item{nb_group}{The number of group.}

\item{nb_elem}{The number of elements to distribute.}

\item{by_row}{A boolean to distribute by row or by column.}
}
\value{
A vector of group indices.
}
\description{
This function distributes elements by group
for a given number of elements. The
distribution can be done by row or by column.
}
\examples{
distribute_by(3, 10)
distribute_by(3, 10, by_row = TRUE)
}
\keyword{internal}
