% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{draw_point}
\alias{draw_point}
\title{Draw arrows}
\usage{
draw_point(
  x,
  y,
  p = NULL,
  ggplot_gen = FALSE,
  cex = par("lwd"),
  pch = 1,
  col = "black"
)
}
\arguments{
\item{p}{ggplot object}

\item{ggplot_gen}{If TRUE add the segments to the ggplot object}

\item{col}{Line color}
}
\value{
Plot the arrows to the current device
or add it to a ggplot object
}
\description{
Draw arrows
}
\keyword{Pedigree-plot}
\keyword{internal}
