% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_observers.R
\name{error_observer}
\alias{error_observer}
\title{Create Error download observers}
\usage{
error_observer(id, df, title)
}
\arguments{
\item{id}{The id of the Shiny module}

\item{df}{A reactive value containing the data to check for errors}

\item{title}{The title of the download button}
}
\description{
Create Error download observers
}
