% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_family_infos.R
\name{family_infos_server}
\alias{family_infos_server}
\title{Server function of selecting columns module}
\usage{
family_infos_server(id, pedi)
}
\arguments{
\item{id}{A string to identify the module.}

\item{df}{A reactive dataframe.}

\item{cols_needed}{A character vector of the mandatory columns.}

\item{cols_supl}{A character vector of the optional columns.}

\item{title}{A string to display in the selectInput.}

\item{na_omit}{A boolean to allow or not the selection of NA.}
}
\value{
A reactive dataframe with the selected columns renamed
to the names of cols_needed and cols_supl.
}
\description{
Server function of selecting columns module
}
\examples{
family_sel_demo()
}
