% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_family_infos.R
\name{family_infos_ui}
\alias{family_infos_ui}
\title{User interface of selecting family module}
\usage{
family_infos_ui(id)
}
\arguments{
\item{id}{A string to identify the module.}
}
\value{
A Shiny module UI.
}
\description{
User interface of selecting family module
}
\examples{
family_sel_demo()
}
