% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_family_sel.R
\name{family_sel_demo}
\alias{family_sel_demo}
\title{Demo function of selecting columns module}
\usage{
family_sel_demo()
}
\description{
Demo function of selecting columns module
}
\examples{
family_sel_demo()
}
