% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{get_families_table}
\alias{get_families_table}
\title{Summarise the families information for a given variable in a data frame}
\usage{
get_families_table(df, var)
}
\arguments{
\item{df}{a data frame}

\item{var}{the variable to summarise}
}
\value{
a data frame with the family information
}
\description{
This function summarises the families information for a given variable in a
data frame. It returns the most numerous modality for each family and the
number of individuals in the family.
}
\examples{
df <- data.frame(
    famid = c(1, 1, 2, 2, 3, 3),
    health = c("A", "B", "A", "A", "B", "B")
)
get_families_table(df, "health")
}
