% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_health_sel.R
\name{health_sel_demo}
\alias{health_sel_demo}
\title{Demo function of selecting health module}
\usage{
health_sel_demo()
}
\description{
Demo function of selecting health module
}
\examples{
health_sel_demo()
}
