% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_health_sel.R
\name{health_sel_server}
\alias{health_sel_server}
\title{Server function of selecting health module}
\usage{
health_sel_server(id, pedi)
}
\arguments{
\item{id}{A string to identify the module.}

\item{df}{A reactive dataframe.}
}
\value{
A reactive dataframe with the selected affection
}
\description{
Server function of selecting health module
}
\examples{
health_sel_demo()
}
