% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_health_sel.R
\name{health_sel_ui}
\alias{health_sel_ui}
\title{User interface of selecting health module}
\usage{
health_sel_ui(id)
}
\arguments{
\item{id}{A string to identify the module.}
}
\value{
A Shiny module UI.
}
\description{
User interface of selecting health module
}
\examples{
health_sel_demo()
}
