% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_inf_sel.R
\name{inf_sel_server}
\alias{inf_sel_server}
\title{Server function of selecting informative module}
\usage{
inf_sel_server(id, pedi)
}
\arguments{
\item{id}{A string to identify the module.}

\item{df}{A reactive dataframe.}
}
\value{
A reactive dataframe with the selected affection
}
\description{
Server function of selecting informative module
}
\examples{
inf_sel_demo()
}
