% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_inf_sel.R
\name{inf_sel_ui}
\alias{inf_sel_ui}
\title{User interface of selecting health module}
\usage{
inf_sel_ui(id)
}
\arguments{
\item{id}{A string to identify the module.}
}
\value{
A Shiny module UI.
}
\description{
User interface of selecting health module
}
\examples{
inf_sel_demo()
}
