% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_class_info}
\alias{make_class_info}
\title{Make class information}
\usage{
make_class_info(x)
}
\arguments{
\item{x}{A list of class}
}
\value{
A character vector of class information
}
\description{
Make class information
}
\examples{
Pedixplorer:::make_class_info(list(1, "a", 1:3, list(1, 2)))
}
\keyword{internal}
