% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_rownames}
\alias{make_rownames}
\title{Make rownames for rectangular data display}
\usage{
make_rownames(x_rownames, nrow, nhead, ntail)
}
\arguments{
\item{x_rownames}{The rownames of the data}

\item{nrow}{The number of rows in the data}

\item{nhead}{The number of rownames to display at the beginning}

\item{ntail}{The number of rownames to display at the end}
}
\value{
A character vector of rownames
}
\description{
Make rownames for rectangular data display
}
\examples{
Pedixplorer::make_rownames(rownames(mtcars), nrow(mtcars), 3, 3)
}
\keyword{internal}
