% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{makeReactive}
\alias{makeReactive}
\title{Make a reactive object}
\usage{
makeReactive(x)
}
\arguments{
\item{x}{The input to check and convert.}
}
\value{
A reactive object.
}
\description{
This function checks if the input is a reactive object. If it is, it returns
it as is. If not, it converts the input into a reactive object.
}
\examples{
Pedixplorer:::makeReactive(1)
}
\keyword{internal}
