% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ped_avaf_infos.R
\name{ped_avaf_infos_server}
\alias{ped_avaf_infos_server}
\alias{ped_avaf_infos_demo}
\title{Shiny modules to display family information}
\usage{
ped_avaf_infos_server(id, pedi)

ped_avaf_infos_demo()
}
\arguments{
\item{id}{A string to identify the module.}

\item{df}{A reactive dataframe.}

\item{cols_needed}{A character vector of the mandatory columns.}

\item{cols_supl}{A character vector of the optional columns.}

\item{title}{A string to display in the selectInput.}

\item{na_omit}{A boolean to allow or not the selection of NA.}
}
\value{
A reactive dataframe with the selected columns renamed
to the names of cols_needed and cols_supl.
}
\description{
This module allows to display the health and availability data
for all individuals in a pedigree object.
The output is a datatable.
The function is composed of two parts: the UI and the server.
The UI is called with the function \code{ped_avaf_infos_ui()} and the server
with the function \code{ped_avaf_infos_server()}.
}
\examples{
family_sel_demo()
}
