% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\docType{data}
\name{ped_ui}
\alias{ped_ui}
\title{Define UI for the Pedigree exploration application}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ped_ui
}
\value{
A shiny UI object
}
\description{
Define UI for the Pedigree exploration application
}
\keyword{datasets}
