% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plink2Pedigree}
\alias{plink2Pedigree}
\title{Import from .fam file or .ped file}
\usage{
plink2Pedigree(
  path,
  sep = "\\t",
  quote = "'",
  header = FALSE,
  na_values = c("NA", "0")
)
}
\arguments{
\item{path}{Path to the file}

\item{sep}{Separator used in the file}
}
\description{
Import a .fam or .ped file and return a Pedigree object
}
