% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_download.R
\name{plot_download_ui}
\alias{plot_download_ui}
\alias{plot_download_server}
\alias{plot_download_demo}
\title{Shiny module to export plot}
\usage{
plot_download_ui(id)

plot_download_server(
  id,
  my_plot,
  filename = "saveplot",
  label = "Download",
  width = 500,
  height = 500,
  ext = "png"
)

plot_download_demo()
}
\arguments{
\item{id}{A string.}

\item{my_plot}{Reactive object containing the plot.}

\item{filename}{A string to name the file.}

\item{label}{A string to name the download button.}

\item{width}{A numeric to set the width of the plot.}

\item{height}{A numeric to set the height of the plot.}

\item{ext}{A string to set the extension of the file.}
}
\description{
This module allow to export multiple type of plot from a reactive object.
The file type currently supported are png, pdf and html.
The function is composed of two parts: the UI and the server.
The UI is called with the function \code{plot_download_ui()} and the server
with the function \code{plot_download_server()}.
}
\examples{
\dontrun{
    plot_download_demo()
}
}
