% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_download.R
\name{plot_download_server}
\alias{plot_download_server}
\title{Export plot server module}
\usage{
plot_download_server(
  id,
  my_plot,
  filename = "saveplot",
  label = "Download",
  width = 500,
  height = 500,
  ext = "png"
)
}
\arguments{
\item{id}{A string.}

\item{my_plot}{Reactive object containing the plot.}
}
\value{
A Shiny UI.
}
\description{
R Shiny module server to export plot
}
\details{
This module allow to export multiple type of plot.
The file type currently supported are png, pdf and html.
The UI ask the user for the file type to export to, the
width and the height of the plot to generate.
When the plot is generated with plotly, set is_plotly to TRUE
and export it to html.
}
