% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_legend.R, R/plot.R
\name{plot_legend_ui}
\alias{plot_legend_ui}
\alias{plot_legend_server}
\alias{plot_legend_demo}
\alias{plot_legend}
\title{Shiny module to generate pedigree graph legend.}
\usage{
plot_legend_ui(id)

plot_legend_server(id, pedi)

plot_legend_demo(max_ind = 500)

plot_legend(
  pedi,
  cex = 1,
  boxw = 0.1,
  boxh = 0.1,
  adjx = 0,
  adjy = 0,
  leg_loc = c(0, 1, 0, 1),
  add_to_existing = FALSE
)
}
\arguments{
\item{id}{A string.}

\item{pedi}{A reactive pedigree object.}
}
\value{
A static UI with the legend.
}
\description{
This module allows to plot the legend of a pedigree object.
The function is composed of two parts: the UI and the server.
The UI is called with the function \code{plot_legend_ui()} and the server
with the function \code{plot_legend_server()}.

Small internal function to be used for plotting a Pedigree
object legend
}
\examples{
if (interactive()) {
    plot_legend_demo()
}
}
\keyword{internal}
\keyword{internal,}
\keyword{plot_legend}
