% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_ped.R
\name{plot_ped_ui}
\alias{plot_ped_ui}
\title{R Shiny module to generate pedigree graph}
\usage{
plot_ped_ui(id)
}
\arguments{
\item{id}{A string.}

\item{df}{A dataframe containing the information about the individual to plot}
}
\value{
A Shiny module.
}
\description{
R Shiny module to generate pedigree graph
}
\examples{
Pedigree_demo()
}
