% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{print_console}
\alias{print_console}
\title{Print to console}
\usage{
print_console(expr, session)
}
\arguments{
\item{expr}{expression to evaluate}

\item{session}{the shiny session}
}
\value{
the result of the expression
}
\description{
This function prints the result of an expression to the console.
Allow easy debugging of the application.
}
