% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ped_avaf_infos.R
\name{sketch}
\alias{sketch}
\title{Sketch of the family information table}
\usage{
sketch(var_name)
}
\arguments{
\item{var_name}{the name of the health variable}
}
\value{
an html sketch of the family information table
}
\description{
Simple function to create a sketch of the family information table.
}
\keyword{internal,}
\keyword{ped_avaf_infos}
