% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fct.R
\name{subregion}
\alias{subregion}
\title{Subset a region of a Pedigree}
\usage{
subregion(df, subreg = NULL)
}
\arguments{
\item{df}{A data frame with all the plot coordinates}

\item{subreg}{A 4-element vector for (min x, max x, min depth, max depth),
used to edit away portions of the plot coordinates returned by
\code{\link[=ped_to_plotdf]{ped_to_plotdf()}}.
This is useful for zooming in on a particular region of the Pedigree.}
}
\value{
A subset of the plot coordinates
}
\description{
Subset a region of a Pedigree
}
\keyword{Pedigree-plot}
\keyword{internal,}
