% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_all.R
\name{plot_all_ui}
\alias{plot_all_ui}
\alias{plot_all_server}
\alias{app_plot_all_demo}
\title{Shiny module with all the components to plot a pedigree}
\usage{
plot_all_ui(id)

plot_all_server(
  id,
  pedi,
  max_ind = 100,
  my_title_l = "My Pedigree",
  my_title_s = "ped_1",
  init_width = "100\%",
  precision = 4
)

app_plot_all_demo()
}
\arguments{
\item{id}{A string to identify the module.}

\item{pedi}{A reactive pedigree object.}

\item{max_ind}{An integer to define the maximum number of individuals
to plot. If the number of individuals is greater than this value,
the user will be asked to confirm the plot.}

\item{my_title_l}{A string to define the title of the plot.}

\item{my_title_s}{A string to define the title of the plot for
the download.}

\item{init_width}{A string to define the initial width of the plot.}

\item{precision}{An integer to define the precision of the plot.}
}
\value{
A reactive list with the plot and the class of the plot.
}
\description{
This module plots a Pedigree object and allows to download
the plot and the data. Different options are available to
customize the plot.
}
\examples{
if (interactive()) {
   app_plot_all_demo()
}
}
