% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data_col_sel.R
\name{data_col_sel_ui}
\alias{data_col_sel_ui}
\alias{data_col_sel_server}
\alias{data_col_sel_demo}
\title{Shiny modules to select columns from a dataframe}
\usage{
data_col_sel_ui(id)

data_col_sel_server(
  id,
  df,
  cols_needed,
  cols_supl,
  title,
  na_omit = TRUE,
  others_cols = TRUE
)

data_col_sel_demo()
}
\arguments{
\item{id}{A string to identify the module.}

\item{df}{A reactive dataframe.}

\item{cols_needed}{A character vector of the mandatory columns.}

\item{cols_supl}{A character vector of the optional columns.}

\item{title}{A string to display in the selectInput.}

\item{na_omit}{A boolean to allow or not the selection of NA.}

\item{others_cols}{A boolean to authorize other columns to be
present in the output datatable.}
}
\value{
A reactive dataframe with the selected columns renamed
to the names of cols_needed and cols_supl.
}
\description{
This function allows to select columns from a dataframe
and rename them to the names of cols_needed and cols_supl.
This generate a Shiny module that can be used in a Shiny app.
The function is composed of two parts: the UI and the server.
The UI is called with the function \code{data_col_sel_ui()} and the server
with the function \code{data_col_sel_server()}.
}
\examples{
\dontrun{
    data_col_sel_demo()
}
}
