% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_family_sel.R, R/app_ped_avaf_infos.R
\name{family_sel_ui}
\alias{family_sel_ui}
\alias{family_sel_server}
\alias{family_sel_demo}
\alias{ped_avaf_infos_ui}
\title{Shiny module to select a family in a pedigree}
\usage{
family_sel_ui(id)

family_sel_server(id, pedi)

family_sel_demo()

ped_avaf_infos_ui(id)
}
\arguments{
\item{id}{A string to identify the module.}

\item{pedi}{A reactive pedigree object.}
}
\value{
A reactive list with the subselected pedigree object and the
selected family id.
}
\description{
This module allows to select a family in a pedigree object.
The function is composed of two parts: the UI and the server.
The UI is called with the function \code{family_sel_ui()} and the server
with the function \code{family_sel_server()}.
}
\examples{
\dontrun{
    family_sel_demo()
}
}
