% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ped_avaf_infos.R
\name{ped_avaf_infos_ui}
\alias{ped_avaf_infos_ui}
\alias{ped_avaf_infos_server}
\alias{ped_avaf_infos_demo}
\title{Shiny modules to display family information}
\usage{
ped_avaf_infos_ui(id, height = "auto")

ped_avaf_infos_server(id, pedi, title = "Family informations", height = "auto")

ped_avaf_infos_demo(height = "auto")
}
\arguments{
\item{id}{A string to identify the module.}

\item{height}{The height of the datatable.}

\item{pedi}{A reactive pedigree object.}

\item{title}{The title of the module.}
}
\value{
A reactive dataframe with the selected columns renamed
to the names of cols_needed and cols_supl.
}
\description{
This module allows to display the health and availability data
for all individuals in a pedigree object.
The output is a datatable.
The function is composed of two parts: the UI and the server.
The UI is called with the function \code{ped_avaf_infos_ui()} and the server
with the function \code{ped_avaf_infos_server()}.
}
\examples{
if (interactive()) {
    ped_avaf_infos_demo()
}
}
\keyword{internal}
