% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{ped_server}
\alias{ped_server}
\title{Create the server logic for the ped_shiny application}
\usage{
ped_server(input, output, session, precision = 6)
}
\arguments{
\item{input}{The input object from a Shiny app.}

\item{output}{The output object from a Shiny app.}

\item{session}{The session object from a Shiny app.}

\item{precision}{Number of decimal for the position of the boxes
in the plot.}
}
\value{
\code{shiny::shinyServer()}
}
\description{
Create the server logic for the ped_shiny application
}
\examples{
if (interactive()) {
    ped_shiny()
}
}
\keyword{internal}
