% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ped_shiny}
\alias{ped_shiny}
\title{Run Pedixplorer Shiny application}
\usage{
ped_shiny(
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1"),
  precision = 2
)
}
\arguments{
\item{port}{(optional) Specify port the application should list to.}

\item{host}{(optional) The IPv4 address that the application should
listen on.}
}
\value{
Running Shiny Application
}
\description{
This function creates a shiny application to manage and visualize
pedigree data using the \code{ped_ui()} and \code{ped_server()} functions.
}
\details{
The application is composed of several modules:
\itemize{
\item Data import
\item Data column selection
\item Data download
\item Family selection
\item Health selection
\item Informative selection
\item Subfamily selection
\item Plotting pedigree
\item Family information
}
}
\examples{
if (interactive()) {
    ped_shiny()
}
}
