% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ped_shiny}
\alias{ped_shiny}
\title{Run Pedixplorer Shiny application}
\usage{
ped_shiny(
  launch_app = TRUE,
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1")
)
}
\arguments{
\item{launch_app}{One of TRUE or FALSE indicating whether or not to run
application. Default is TRUE.}

\item{port}{(optional) Specify port the application should list to.}

\item{host}{(optional) The IPv4 address that the application should
listen on.}
}
\value{
Running Shiny Application
}
\description{
Main function to run Pedixplorer Shiny application.
}
\examples{
\dontrun{
## To run app set launch_app=TRUE
ped_shiny(launch_app=FALSE)
}
}
