% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/app_server.R, R/app_ui.R
\docType{data}
\name{ped_shiny}
\alias{ped_shiny}
\alias{ped_server}
\alias{ped_ui}
\title{Run Pedixplorer Shiny application}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ped_shiny(
  launch_app = TRUE,
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1")
)

ped_server(input, output, session)

ped_ui
}
\arguments{
\item{launch_app}{One of TRUE or FALSE indicating whether or not to run
application. Default is TRUE.}

\item{port}{(optional) Specify port the application should list to.}

\item{host}{(optional) The IPv4 address that the application should
listen on.}

\item{input}{The input object from a Shiny app.}

\item{output}{The output object from a Shiny app.}

\item{session}{The session object from a Shiny app.}
}
\value{
Running Shiny Application
}
\description{
This function creates a shiny application to manage and visualize
pedigree data using the \code{ped_ui()} and \code{ped_server()} functions.
}
\details{
The application is composed of several modules:
\itemize{
\item Data import
\item Data column selection
\item Data download
\item Family selection
\item Health selection
\item Informative selection
\item Subfamily selection
\item Plotting pedigree
\item Family information
}
}
\examples{
\dontrun{
## To run app set launch_app=TRUE
ped_shiny(launch_app=FALSE)
}
}
\keyword{datasets}
