% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{ped_ui}
\alias{ped_ui}
\title{Create the user interface for the ped_shiny application}
\value{
\code{shiny::shinyUI()}
}
\description{
Create the user interface for the ped_shiny application
}
\examples{
if (interactive()) {
    ped_shiny()
}
}
\keyword{internal}
