% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pedigree.R
\name{plot_legend}
\alias{plot_legend}
\title{Plot legend}
\usage{
plot_legend(
  obj,
  cex = 1,
  boxw = 0.1,
  boxh = 0.1,
  adjx = 0,
  adjy = 0,
  leg_loc = c(0, 1, 0, 1),
  add_to_existing = FALSE,
  usr = NULL,
  lwd = par("lwd"),
  precision = 4
)
}
\arguments{
\item{obj}{A Pedigree object}

\item{cex}{Character expansion of the text}

\item{boxw}{Width of the polygons elements}

\item{boxh}{Height of the polygons elements}

\item{adjx}{default=0.  Controls the horizontal text adjustment of
the labels in the legend.}

\item{adjy}{default=0.  Controls the vertical text adjustment
of the labels in the legend.}

\item{lwd}{default=par("lwd").  Controls the bordering line width of the
elements in the legend.}
}
\value{
an invisible list containing
\itemize{
\item df : the data.frame used to plot the Pedigree
\item par_usr : the user coordinates used to plot the Pedigree
}
}
\description{
Small internal function to be used for plotting a Pedigree
object legend
}
\section{Side Effects}{

Creates plot on current plotting device.
}

\keyword{internal}
\keyword{plot_legend}
