% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_ped.R
\name{plot_ped_ui}
\alias{plot_ped_ui}
\alias{plot_ped_server}
\alias{plot_ped_demo}
\title{Shiny module to generate pedigree graph.}
\usage{
plot_ped_ui(id)

plot_ped_server(
  id,
  pedi,
  my_title = NA,
  precision = 2,
  my_tips = NULL,
  plot_lwd = 1,
  width = "80\%",
  height = "400px",
  plot_cex = 1,
  symbolsize = 1,
  force = TRUE,
  plot_par = list(),
  is_interactive = FALSE,
  aff_mark = TRUE,
  label = NULL
)

plot_ped_demo(pedi = NULL, precision = 4, interactive = FALSE)
}
\arguments{
\item{id}{A string.}

\item{pedi}{A reactive pedigree object.}

\item{precision}{An integer to set the precision of the plot.}

\item{title}{A string to name the plot.}

\item{max_ind}{An integer to set the maximum number of individuals to plot.}

\item{tips}{A character vector of the column names of the data frame to use
as tooltips. If NULL, no tooltips are added.}

\item{lwd}{A numeric to set the line width of the plot.}
}
\value{
A reactive ggplot or the pedigree object.
}
\description{
This module allows to plot a pedigree object. The plot can be interactive.
The function is composed of two parts: the UI and the server.
The UI is called with the function \code{plot_ped_ui()} and the server
with the function \code{plot_ped_server()}.
}
\examples{
if (interactive()) {
    data("sampleped")
    pedi <- shiny::reactive({
        Pedigree(sampleped[sampleped$famid == "1", ])
    })
    plot_ped_demo(pedi)
}
}
\keyword{internal}
