% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_ped.R
\name{plot_ped_ui}
\alias{plot_ped_ui}
\alias{plot_ped_server}
\alias{plot_ped_demo}
\title{Shiny module to generate pedigree graph.}
\usage{
plot_ped_ui(id)

plot_ped_server(id, pedi, title, max_ind = 500)

plot_ped_demo(pedi, max_ind = 500)
}
\arguments{
\item{id}{A string.}

\item{pedi}{A reactive pedigree object.}

\item{title}{A string to name the plot.}
}
\value{
A reactive ggplot or the pedigree object.
}
\description{
This module allows to plot a pedigree object. The plot can be interactive.
The function is composed of two parts: the UI and the server.
The UI is called with the function \code{plot_ped_ui()} and the server
with the function \code{plot_ped_server()}.
}
\examples{
if (interactive()) {
    data("sampleped")
    pedi <- shiny::reactive({
        Pedigree(sampleped[sampleped$famid == "1", ])
    })
    plot_ped_demo(pedi)
}
}
\keyword{internal}
