% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_ped.R
\name{plot_ped_ui}
\alias{plot_ped_ui}
\alias{plot_ped_server}
\alias{plot_ped_demo}
\title{Shiny module to generate pedigree graph.}
\usage{
plot_ped_ui(id)

plot_ped_server(
  id,
  pedi,
  title,
  precision = 2,
  max_ind = 500,
  lwd = par("lwd"),
  tips = NULL
)

plot_ped_demo(pedi, precision = 2, max_ind = 500, tips = NULL)
}
\arguments{
\item{id}{A string.}

\item{pedi}{A reactive pedigree object.}

\item{title}{A string to name the plot.}

\item{precision}{An integer to set the precision of the plot.}

\item{max_ind}{An integer to set the maximum number of individuals to plot.}

\item{lwd}{default=par("lwd"). Controls the line width of the
segments, arcs and polygons.}

\item{tips}{A character vector of the column names of the data frame to
use as tooltips. If \code{NULL}, no tooltips are added.}
}
\value{
A reactive ggplot or the pedigree object.
}
\description{
This module allows to plot a pedigree object. The plot can be interactive.
The function is composed of two parts: the UI and the server.
The UI is called with the function \code{plot_ped_ui()} and the server
with the function \code{plot_ped_server()}.
}
\examples{
if (interactive()) {
    data("sampleped")
    pedi <- shiny::reactive({
        Pedigree(sampleped[sampleped$famid == "1", ])
    })
    plot_ped_demo(pedi)
}
}
\keyword{internal}
